refactor: Refactor appointment form to utilize patient details from AuthService
